fill chatDB with the corresponding data and delete temp file
