+++ src/data/chatDB.ts
@@ -1,158 +1,1276 @@
-// Mock chat database
-let chatsDB = [
+import type { chatsDBProps } from "../types/chat";
+
+export const chatsDB: chatsDBProps[] = [
   {
     id: 1,
-    name: "María García",
-    avatar:
-      "https://images.pexels.com/photos/774909/pexels-photo-774909.jpeg?auto=compress&cs=tinysrgb&w=100",
-    lastMessage: "¡Hola! ¿Cómo estás?",
-    lastMessageTime: new Date(Date.now() - 5 * 60 * 1000), // 5 minutes ago
-    unreadCount: 2,
+    name: "Pablo Demarchi",
+    avatar: "/Avatars/image_0000.webp",
+    lastMessage: "",
+    lastMessageTime: new Date(Date.now() - 5 * 60 * 1000),
+    unreadCount: 0,
     isOnline: true,
     messages: [
       {
-        id: 1,
-        text: "¡Hola! ¿Cómo estás?",
-        timestamp: new Date(Date.now() - 10 * 60 * 1000),
+        text: "Che al final encontraste el lugar para estacionar?",
+        timestamp: "6:46 p.m.",
         sender: "contact",
         status: "read",
+        id: 1,
       },
       {
-        id: 2,
-        text: "Todo bien, ¿y tú?",
-        timestamp: new Date(Date.now() - 8 * 60 * 1000),
+        text: "Sí, me costó un poco pero dejé el auto a la vuelta jeje, por suerte nadie me vió estacionando 🤣🤣🤣🤣🤣. Tuve que hacer malabares para meter el auto en ese lugarcito que dejaron jajaja",
+        timestamp: "6:47 p.m.",
         sender: "me",
         status: "read",
+        id: 2,
       },
       {
-        id: 3,
-        text: "Perfecto! ¿Nos vemos mañana?",
-        timestamp: new Date(Date.now() - 5 * 60 * 1000),
+        text: "Bueno joya, ya entro entonces!",
+        timestamp: "6:48 p.m.",
         sender: "contact",
-        status: "delivered",
+        status: "read",
+        id: 3,
+      },
+      {
+        text: "Tremenda la peli eh, la verdad me re sorprendió",
+        timestamp: "9:30 p.m.",
+        sender: "me",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "Mal! La escena donde pelean en el agua me dejó loco 😱, estaban muy buenos los efectos la verdad y encima el sonido era brutal, aunque por ahí se saturaba un poco nomás 😂😂😂😂😂, pero todo lo demas estubo súper",
+        timestamp: "9:31 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Siii, no me la esperaba ni ahí",
+        timestamp: "9:32 p.m.",
+        sender: "me",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "Después te paso un video con teorías que vi en YouTube jajaj",
+        timestamp: "9:33 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Dale, estaría genial eso. Estoy re manija jaja",
+        timestamp: "9:34 p.m.",
+        sender: "me",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "Ahora justamente estaba viendo a un YouTuber que hablaba de la peli",
+        timestamp: "9:34 p.m.",
+        sender: "me",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "No le entendí mucho así que me voy a comer algo y después sigo",
+        timestamp: "9:34 p.m.",
+        sender: "me",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Che y mañana qué onda? Vas a lo de Nico?",
+        timestamp: "10:02 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "No sé todavía, me siento medio bajón hoy la verdad",
+        timestamp: "10:04 p.m.",
+        sender: "me",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "Uh amigo, todo bien?",
+        timestamp: "10:05 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "Sí, tranqui, solo depre de domingo jaja. Viste cómo es.",
+        timestamp: "10:06 p.m.",
+        sender: "me",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Te entiendo. Si querés venite a casa y comemos algo, sin plan ni nada.",
+        timestamp: "10:07 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "Re va eso, gracias loco ❤️",
+        timestamp: "10:08 p.m.",
+        sender: "me",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "Para eso estamos pa",
+        timestamp: "10:08 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 17,
+      },
+      {
+        text: "Mañana te caigo tipo 8 entonces",
+        timestamp: "10:10 p.m.",
+        sender: "me",
+        status: "read",
+        id: 18,
+      },
+      {
+        text: "Dale, te espero con empanadas 😎",
+        timestamp: "10:11 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 19,
+      },
+      {
+        text: "Jajaja sos un grande",
+        timestamp: "10:12 p.m.",
+        sender: "me",
+        status: "read",
+        id: 20,
+      },
+      {
+        text: "Che viste lo de la tormenta?",
+        timestamp: "8:22 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 21,
+      },
+      {
+        text: "Sí, me despertó el viento 😵",
+        timestamp: "8:23 a.m.",
+        sender: "me",
+        status: "read",
+        id: 22,
+      },
+      {
+        text: "Encima se cortó la luz un toque acá",
+        timestamp: "8:24 a.m.",
+        sender: "me",
+        status: "read",
+        id: 23,
+      },
+      {
+        text: "Acá también. No puedo prepararme ni el café, encima me levante temprano para terminar de hacer un trabajo. y bueno... lo dejaré para después de comer 🤣🤣🤣",
+        timestamp: "8:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 24,
+      },
+      {
+        text: "Iagual, creo que es mejor descansar y esperar a que pase todo",
+        timestamp: "8:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 25,
+      },
+      {
+        text: "Por las dudas voy a desenchufar mis cosas jajaja",
+        timestamp: "8:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 26,
+      },
+      {
+        text: "Drama real jaja",
+        timestamp: "8:25 a.m.",
+        sender: "me",
+        status: "read",
+        id: 27,
+      },
+      {
+        text: "Creo que yo debería hacer lo mismo",
+        timestamp: "8:26 a.m.",
+        sender: "me",
+        status: "read",
+        id: 28,
+      },
+      {
+        text: "🤣🤣🤣🤣🤣",
+        timestamp: "8:26 a.m.",
+        sender: "me",
+        status: "read",
+        id: 29,
+      },
+      {
+        text: "Che me olvidé de contarte, me llamaron de laburo nuevo",
+        timestamp: "3:00 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 30,
+      },
+      {
+        text: "Ehhh qué bien amigooo! Te felicitoooo 🥳, tenemos que salir a festejar en algún lugar amigo, encima con la dedicación que tenés vas a llegar a lejos amigo!!! Ahora le voy a decir a mamá que se va a poner súper contenta",
+        timestamp: "3:01 p.m.",
+        sender: "me",
+        status: "read",
+        id: 31,
+      },
+      {
+        text: "Gracias loco, todavía no lo puedo creer jajaja",
+        timestamp: "3:02 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 32,
+      },
+      {
+        text: "Te re lo merecés, posta",
+        timestamp: "3:03 p.m.",
+        sender: "me",
+        status: "read",
+        id: 33,
+      },
+      {
+        text: "Vamos a tener que festejar eso eh",
+        timestamp: "3:04 p.m.",
+        sender: "me",
+        status: "read",
+        id: 34,
+      },
+      {
+        text: "Ni hablar. Este finde armamos algo.",
+        timestamp: "3:05 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 35,
+      },
+      {
+        text: "Avísame y llevo birras 😏",
+        timestamp: "3:06 p.m.",
+        sender: "me",
+        status: "read",
+        id: 36,
       },
     ],
   },
   {
     id: 2,
-    name: "Carlos Rodríguez",
-    avatar:
-      "https://images.pexels.com/photos/220453/pexels-photo-220453.jpeg?auto=compress&cs=tinysrgb&w=100",
-    lastMessage: "La reunión es a las 3pm",
-    lastMessageTime: new Date(Date.now() - 30 * 60 * 1000), // 30 minutes ago
+    name: "Martín Pérez",
+    avatar: "/Avatars/image_0002.webp",
+    lastMessage: "Sí, de una, que venga si jajajaja.",
+    lastMessageTime: new Date(Date.now() - 30 * 60 * 1000), // 30 minutes ago,
     unreadCount: 0,
     isOnline: false,
     lastSeen: new Date(Date.now() - 15 * 60 * 1000),
     messages: [
       {
-        id: 1,
-        text: "¿A qué hora es la reunión?",
-        timestamp: new Date(Date.now() - 45 * 60 * 1000),
-        sender: "me",
-        status: "read",
-      },
-      {
-        id: 2,
-        text: "La reunión es a las 3pm",
-        timestamp: new Date(Date.now() - 30 * 60 * 1000),
+        text: "Che, al final vas a la juntada el sábado?",
+        timestamp: "10:01 a.m.",
         sender: "contact",
         status: "read",
+        id: 1,
+      },
+      {
+        text: "Mirá que no podes faltar amigo vos he jajaja, hace mucho no nos vemos así que nos viene de 10",
+        timestamp: "10:01 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 2,
+      },
+      {
+        text: "Tenemos que ponernos al día con nuestras cosas 🤣🤣🤣🤣",
+        timestamp: "10:01 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 3,
+      },
+      {
+        text: "Sí, obvio. A qué hora era?",
+        timestamp: "10:02 a.m.",
+        sender: "me",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "Ya le avisé a la Eli que si o si voy jajaja asi que nos vemos allá seguro amigo",
+        timestamp: "10:02 a.m.",
+        sender: "me",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Avisame nomas la hora y caigo 🚀",
+        timestamp: "10:02 a.m.",
+        sender: "me",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "Tipo 8 voy maso, pero podés caer antes si querés así habalmos.",
+        timestamp: "10:03 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Dale, capaz llego 7 y media así ayudo con algo.",
+        timestamp: "10:04 a.m.",
+        sender: "me",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "Buenísimo, traé la Play si podés de paso 🤣🤣🤣🤣.",
+        timestamp: "10:05 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "La mia dejó de andar hace unos 2 mesesmaso 😑😑😑😑",
+        timestamp: "10:05 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Obvio, llevo los joysticks también.",
+        timestamp: "10:06 a.m.",
+        sender: "me",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "Tengo un juego de cartas que está buenisimo, ese tambien voy a tratar de llevar",
+        timestamp: "10:06 a.m.",
+        sender: "me",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "Voy a pedirle a la Eli que lleve alargue para enchufar todo nomás 🤣🤣🤣🤣",
+        timestamp: "10:06 a.m.",
+        sender: "me",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "Che, quién más va?",
+        timestamp: "10:07 a.m.",
+        sender: "me",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Dalee perfecto, me parece genial jajajaja. Viene el Nico, la Vale, el Fede y capaz la Sofi.",
+        timestamp: "10:08 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "Despues otros amigos de Eli que no se quienes son jajajaja",
+        timestamp: "10:08 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "Buee la Sofiii. Hace mil que no la veo jajaja.",
+        timestamp: "10:09 a.m.",
+        sender: "me",
+        status: "read",
+        id: 17,
+      },
+      {
+        text: "Sí, encima dice que anda con ganas de salir jajajaja 🤣🤣.",
+        timestamp: "10:10 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 18,
+      },
+      {
+        text: "Che, y no sabes si hay que llevar algo mas para tomar?",
+        timestamp: "10:11 a.m.",
+        sender: "me",
+        status: "read",
+        id: 19,
+      },
+      {
+        text: "Lo que vos quieras tomas nomás, acá creeria que va a haber algo, pero no sé que.",
+        timestamp: "10:12 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 20,
+      },
+      {
+        text: "Dale, paso por el chino antes de ir entonces y compro.",
+        timestamp: "10:13 a.m.",
+        sender: "me",
+        status: "read",
+        id: 21,
+      },
+      {
+        text: "Hace falta hielo tambien?",
+        timestamp: "10:14 a.m.",
+        sender: "me",
+        status: "read",
+        id: 22,
+      },
+      {
+        text: "Tienen muy buenas ofertas ahora así que llevo",
+        timestamp: "10:14 a.m.",
+        sender: "me",
+        status: "read",
+        id: 23,
+      },
+      {
+        text: "Sí, si querés traete un par de bolsas por las dudas.",
+        timestamp: "10:15 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 24,
+      },
+      {
+        text: "Listo, anoto todo. Algo más?",
+        timestamp: "10:16 a.m.",
+        sender: "me",
+        status: "read",
+        id: 25,
+      },
+      {
+        text: "No, yo creo que con eso estamos. Nos vemos el sábado amigazo!",
+        timestamp: "10:17 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 26,
+      },
+      {
+        text: "Dale mi hermano, cualquier cosa te estoy chiflando",
+        timestamp: "10:18 a.m.",
+        sender: "me",
+        status: "read",
+        id: 27,
+      },
+      {
+        text: "Espero no me pase lo de la play 🤣🤣🤣🤣",
+        timestamp: "10:18 a.m.",
+        sender: "me",
+        status: "read",
+        id: 28,
+      },
+      {
+        text: "Mientras tanto voy preparando el juego de cartas, está buenísimo, tenes que formar dos grupos y vos tenés que tratar de hacer que tu equipo adivine la mayor cantidad de cartas en el menor tiempo posible. En la carta, abajo aparecen palabras que no podes decir. Está muy bueno!!",
+        timestamp: "10:18 a.m.",
+        sender: "me",
+        status: "read",
+        id: 29,
+      },
+      {
+        text: "Que copado amigo, suena re interesante jajaja. Che tenés la direccion del lugar? sino te paso.",
+        timestamp: "10:19 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 30,
+      },
+      {
+        text: "Sí, la tengo pero pasala igual así la guardo.",
+        timestamp: "10:20 a.m.",
+        sender: "me",
+        status: "read",
+        id: 31,
+      },
+      {
+        text: "Es en la casa de siempre, la de la esquina con el portón verde.",
+        timestamp: "10:21 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 32,
+      },
+      {
+        text: "Perfecto, ya sé cuál es 🙂.",
+        timestamp: "10:22 a.m.",
+        sender: "me",
+        status: "read",
+        id: 33,
+      },
+      {
+        text: "Vas a venir con alguien?",
+        timestamp: "10:23 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 34,
+      },
+      {
+        text: "Capaz vaya con mi primo Ale,",
+        timestamp: "10:24 a.m.",
+        sender: "me",
+        status: "read",
+        id: 35,
+      },
+      {
+        text: "Sí, de una, que venga si jajajaja.",
+        timestamp: "10:25 a.m.",
+        sender: "contact",
+        status: "delivered",
+        id: 36,
       },
     ],
   },
   {
     id: 3,
-    name: "Ana López",
-    avatar:
-      "https://images.pexels.com/photos/415829/pexels-photo-415829.jpeg?auto=compress&cs=tinysrgb&w=100",
-    lastMessage: "¡Gracias por tu ayuda!",
+    name: "Andrea Gómez",
+    avatar: "/Avatars/image_0003.webp",
+    lastMessage: "No, tranqui, voy directo. Nos vemos allá!",
     lastMessageTime: new Date(Date.now() - 2 * 60 * 60 * 1000), // 2 hours ago
     unreadCount: 1,
     isOnline: true,
     messages: [
       {
-        id: 1,
-        text: "¿Me puedes ayudar con el proyecto?",
-        timestamp: new Date(Date.now() - 3 * 60 * 60 * 1000),
+        text: "Nos vemos en el café tipo 5?",
+        timestamp: "3:10 p.m.",
         sender: "contact",
         status: "read",
+        id: 1,
       },
       {
-        id: 2,
-        text: "¡Por supuesto! ¿Qué necesitas?",
-        timestamp: new Date(Date.now() - 2.5 * 60 * 60 * 1000),
+        text: "Dale, ahí nos vemos.",
+        timestamp: "3:12 p.m.",
         sender: "me",
         status: "read",
+        id: 2,
       },
       {
+        text: "¿Querés que lleve algo?",
+        timestamp: "3:13 p.m.",
+        sender: "me",
+        status: "read",
         id: 3,
-        text: "¡Gracias por tu ayuda!",
-        timestamp: new Date(Date.now() - 2 * 60 * 60 * 1000),
+      },
+      {
+        text: "No hace falta, tranqui.",
+        timestamp: "3:14 p.m.",
         sender: "contact",
-        status: "delivered",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "Bueno, igual llevo unas medialunas por las dudas jaja",
+        timestamp: "3:15 p.m.",
+        sender: "me",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Jajaja, sos un crack",
+        timestamp: "3:16 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "¿Te acordás dónde queda el café, no?",
+        timestamp: "3:17 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Sí, el de siempre, ¿no?",
+        timestamp: "3:18 p.m.",
+        sender: "me",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "Ese mismo. Igual te paso la ubicación por las dudas.",
+        timestamp: "3:19 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "Dale, mandala así la guardo.",
+        timestamp: "3:20 p.m.",
+        sender: "me",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Listo, te la mandé por WhatsApp.",
+        timestamp: "3:21 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "Perfecto, ya la tengo.",
+        timestamp: "3:22 p.m.",
+        sender: "me",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "¿Vas a ir en auto o en bondi?",
+        timestamp: "3:23 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "Voy en auto, así después te llevo si querés.",
+        timestamp: "3:24 p.m.",
+        sender: "me",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Buenísimo, gracias!",
+        timestamp: "3:25 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "¿Querés que pase a buscarte?",
+        timestamp: "3:26 p.m.",
+        sender: "me",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "No, tranqui, voy directo. Nos vemos allá!",
+        timestamp: "3:27 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 17,
       },
     ],
   },
   {
     id: 4,
-    name: "Grupo Familia",
-    avatar:
-      "https://images.pexels.com/photos/1128318/pexels-photo-1128318.jpeg?auto=compress&cs=tinysrgb&w=100",
-    lastMessage: "Mamá: ¿Vienen a cenar el domingo?",
+    name: "Alberto Torres",
+    avatar: "/Avatars/image_0004.webp",
+    lastMessage: "Obvio, te mando foto.",
     lastMessageTime: new Date(Date.now() - 4 * 60 * 60 * 1000), // 4 hours ago
-    unreadCount: 5,
-    isGroup: true,
+    unreadCount: 0,
     isOnline: false,
     messages: [
       {
-        id: 1,
-        text: "¡Hola familia!",
-        timestamp: new Date(Date.now() - 5 * 60 * 60 * 1000),
+        text: "Te viste el último capi de la serie?",
+        timestamp: "7:15 p.m.",
         sender: "contact",
-        senderName: "Papá",
         status: "read",
+        id: 1,
       },
       {
-        id: 2,
-        text: "¡Hola papá!",
-        timestamp: new Date(Date.now() - 4.5 * 60 * 60 * 1000),
+        text: "Sí! Altísimo final.",
+        timestamp: "7:20 p.m.",
         sender: "me",
         status: "read",
+        id: 2,
       },
       {
+        text: "No me esperaba ese giro, posta.",
+        timestamp: "7:21 p.m.",
+        sender: "me",
+        status: "read",
         id: 3,
-        text: "¿Vienen a cenar el domingo?",
-        timestamp: new Date(Date.now() - 4 * 60 * 60 * 1000),
+      },
+      {
+        text: "Yo tampoco, me dejó re manija.",
+        timestamp: "7:22 p.m.",
         sender: "contact",
-        senderName: "Mamá",
-        status: "delivered",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "¿Cuándo sale el próximo?",
+        timestamp: "7:23 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Creo que el viernes que viene.",
+        timestamp: "7:24 p.m.",
+        sender: "me",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "Uff, falta una banda.",
+        timestamp: "7:25 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Encima vi teorías re locas en Twitter.",
+        timestamp: "7:26 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "Pasame alguna, así me cebás más.",
+        timestamp: "7:27 p.m.",
+        sender: "me",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "Después te paso el hilo, está buenísimo.",
+        timestamp: "7:28 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Dale, lo leo después de cenar.",
+        timestamp: "7:29 p.m.",
+        sender: "me",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "¿Ya comiste?",
+        timestamp: "7:30 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "Todavía no, estoy viendo qué hago.",
+        timestamp: "7:31 p.m.",
+        sender: "me",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "Pedite una pizza, es la que va.",
+        timestamp: "7:32 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Me tentaste, ahora llamo a la pizzería.",
+        timestamp: "7:33 p.m.",
+        sender: "me",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "Jajaja, después contame si estaba buena.",
+        timestamp: "7:34 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "Obvio, te mando foto.",
+        timestamp: "7:35 p.m.",
+        sender: "me",
+        status: "read",
+        id: 17,
       },
     ],
   },
   {
     id: 5,
-    name: "David Martín",
-    avatar:
-      "https://images.pexels.com/photos/91227/pexels-photo-91227.jpeg?auto=compress&cs=tinysrgb&w=100",
-    lastMessage: "¡Perfecto! Nos vemos entonces",
+    name: "Franco Fernández",
+    avatar: "/Avatars/image_0005.webp",
+    lastMessage: "Cualquier cosa avisame, estoy en la compu.",
     lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
     unreadCount: 0,
     isOnline: false,
     lastSeen: new Date(Date.now() - 12 * 60 * 60 * 1000),
     messages: [
       {
-        id: 1,
-        text: "¿Quieres ir al cine mañana?",
-        timestamp: new Date(Date.now() - 25 * 60 * 60 * 1000),
-        sender: "me",
-        status: "read",
-      },
-      {
-        id: 2,
-        text: "¡Perfecto! Nos vemos entonces",
-        timestamp: new Date(Date.now() - 24 * 60 * 60 * 1000),
+        text: "Che, necesito una mano con el proyecto.",
+        timestamp: "11:00 a.m.",
         sender: "contact",
         status: "read",
+        id: 1,
+      },
+      {
+        text: "Decime nomás, en qué te doy una mano?",
+        timestamp: "11:02 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+      {
+        text: "No entiendo bien cómo hacer el login.",
+        timestamp: "11:03 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 3,
+      },
+      {
+        text: "¿Tenés el repo a mano?",
+        timestamp: "11:04 a.m.",
+        sender: "me",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "Sí, te lo paso por mail.",
+        timestamp: "11:05 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Dale, avisame cuando lo mandes.",
+        timestamp: "11:06 a.m.",
+        sender: "me",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "Listo, ya te lo mandé.",
+        timestamp: "11:07 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Perfecto, ahora lo miro.",
+        timestamp: "11:08 a.m.",
+        sender: "me",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "¿Te paso mi usuario y clave de prueba?",
+        timestamp: "11:09 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "Sí, así pruebo bien.",
+        timestamp: "11:10 a.m.",
+        sender: "me",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Usuario: franco, clave: 1234",
+        timestamp: "11:11 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "Listo, ya entré.",
+        timestamp: "11:12 a.m.",
+        sender: "me",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "¿Te tira algún error?",
+        timestamp: "11:13 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "No, pero el botón de login no se desactiva.",
+        timestamp: "11:14 a.m.",
+        sender: "me",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Ah, seguro es por el estado. Fijate si lo actualizás bien.",
+        timestamp: "11:15 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "Dale, ahora lo reviso.",
+        timestamp: "11:16 a.m.",
+        sender: "me",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "Cualquier cosa avisame, estoy en la compu.",
+        timestamp: "11:17 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 17,
+      },
+    ],
+  },
+  {
+    id: 6,
+    name: "Jorge Martínez",
+    avatar: "/Avatars/image_0006.webp",
+    lastMessage: "Gracias, igualmente!",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: true,
+    messages: [
+      {
+        text: "Hola Jorge, me pasás el apunte de historia?",
+        timestamp: "8:30 a.m.",
+        sender: "me",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Sí, bancá que lo busco.",
+        timestamp: "8:31 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 2,
+      },
+      {
+        text: "Gracias! Me re salvas.",
+        timestamp: "8:32 a.m.",
+        sender: "me",
+        status: "read",
+        id: 3,
+      },
+      {
+        text: "No hay drama, para eso estamos.",
+        timestamp: "8:33 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 4,
+      },
+      {
+        text: "Te lo mando por mail o por acá?",
+        timestamp: "8:34 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 5,
+      },
+      {
+        text: "Por acá mejor, así lo bajo al toque.",
+        timestamp: "8:35 a.m.",
+        sender: "me",
+        status: "read",
+        id: 6,
+      },
+      {
+        text: "Listo, ahí va el PDF.",
+        timestamp: "8:36 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 7,
+      },
+      {
+        text: "Llegó perfecto, mil gracias.",
+        timestamp: "8:37 a.m.",
+        sender: "me",
+        status: "read",
+        id: 8,
+      },
+      {
+        text: "Ya estudiaste algo?",
+        timestamp: "8:38 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 9,
+      },
+      {
+        text: "Un poco, pero me falta banda todavía.",
+        timestamp: "8:39 a.m.",
+        sender: "me",
+        status: "read",
+        id: 10,
+      },
+      {
+        text: "Tranqui, yo tampoco arranqué fuerte.",
+        timestamp: "8:40 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 11,
+      },
+      {
+        text: "Querés que repasemos juntos después?",
+        timestamp: "8:41 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 12,
+      },
+      {
+        text: "Sí, de una. Hoy a la tarde podés?",
+        timestamp: "8:42 a.m.",
+        sender: "me",
+        status: "read",
+        id: 13,
+      },
+      {
+        text: "Dale, tipo 5 te va?",
+        timestamp: "8:43 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 14,
+      },
+      {
+        text: "Perfecto, te escribo cuando esté libre.",
+        timestamp: "8:44 a.m.",
+        sender: "me",
+        status: "read",
+        id: 15,
+      },
+      {
+        text: "Dale, suerte con el estudio!",
+        timestamp: "8:45 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 16,
+      },
+      {
+        text: "Gracias, igualmente!",
+        timestamp: "8:46 a.m.",
+        sender: "me",
+        status: "read",
+        id: 17,
+      },
+    ],
+  },
+  {
+    id: 7,
+    name: "Sofía Ramírez",
+    avatar: "/Avatars/image_0007.webp",
+    lastMessage: "Gracias genia! 🥳",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: true,
+    messages: [
+      {
+        text: "Feliz cumple 🎉",
+        timestamp: "7:55 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Gracias genia! 🥳",
+        timestamp: "7:58 p.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 8,
+    name: "Marco Perdomo",
+    avatar: "/Avatars/image_0008.webp",
+    lastMessage: "Ey Marco, puedo pasar por tu casa a buscar los apuntes?",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Buen día, revisaste el informe?",
+        timestamp: "9:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Ey Marco, puedo pasar por tu casa a buscar los apuntes?",
+        timestamp: "9:28 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 9,
+    name: "Paula Lezcano",
+    avatar: "/Avatars/image_0009.webp",
+    lastMessage: "Hola Pau, sigue estando para pasar a buscarlo?",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Buen día, miraste el informe?",
+        timestamp: "9:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Hola Pau, sigue estando para pasar a buscarlo?",
+        timestamp: "9:28 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 10,
+    name: "Pedro Falcón",
+    avatar: "/Avatars/image_0010.webp",
+    lastMessage: "Todavía no pude, pero mañana sin falta te lo paso.",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Buen día, revisaste el informe?",
+        timestamp: "9:25 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Todavía no pude, pero mañana sin falta te lo paso.",
+        timestamp: "9:28 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 11,
+    name: "Lucía Méndez",
+    avatar: "/Avatars/image_0011.webp",
+    lastMessage: "Sí, todo perfecto. Gracias che.",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: true,
+    messages: [
+      {
+        text: "Te llegó el paquete al final?",
+        timestamp: "8:00 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Sí, todo perfecto. Gracias che.",
+        timestamp: "8:05 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 12,
+    name: "Diego Rivera",
+    avatar: "/Avatars/image_0012.webp",
+    lastMessage: "Dale! Voy a estar al toque.",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Tenemos la reunión a las 3 eh.",
+        timestamp: "11:15 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Dale! Voy a estar al toque.",
+        timestamp: "11:17 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 13,
+    name: "Valeria Suárez",
+    avatar: "/Avatars/image_0013.webp",
+    lastMessage: "Obvio! Te lo mando ya.",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Me pasás el archivo?",
+        timestamp: "10:20 a.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Obvio! Te lo mando ya.",
+        timestamp: "10:21 a.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 14,
+    name: "Camila Quintana",
+    avatar: "/Avatars/image_0014.webp",
+    lastMessage: "Dale! Después de cenar me conecto.",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Jugamos esta noche?",
+        timestamp: "6:30 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Dale! Después de cenar me conecto.",
+        timestamp: "6:33 p.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
+      },
+    ],
+  },
+  {
+    id: 15,
+    name: "Matías Ortega",
+    avatar: "/Avatars/image_0015.webp",
+    lastMessage: "Sí, ya les respondí. Vos qué onda?",
+    lastMessageTime: new Date(Date.now() - 24 * 60 * 60 * 1000), // 1 day ago
+    unreadCount: 0,
+    isOnline: false,
+    messages: [
+      {
+        text: "Viste el correo de RRHH?",
+        timestamp: "4:45 p.m.",
+        sender: "contact",
+        status: "read",
+        id: 1,
+      },
+      {
+        text: "Sí, ya les respondí. Vos qué onda?",
+        timestamp: "4:47 p.m.",
+        sender: "me",
+        status: "read",
+        id: 2,
       },
     ],
   },
@@ -164,10 +1282,8 @@ export const getNextChatId = () => {
 };
 
 // Get next available ID for messages
-export const getNextMessageId = (chatId) => {
+export const getNextMessageId = (chatId: number) => {
   const chat = chatsDB.find((c) => c.id === chatId);
   if (!chat || !chat.messages.length) return 1;
   return Math.max(...chat.messages.map((msg) => msg.id)) + 1;
 };
-
-export { chatsDB };
