--- src/data/temp.ts
@@ -1,1292 +1,0 @@
-export type ChatMessage = {
-  id: number;
-  text: string;
-  timestamp: Date | string;
-  sender: "contact" | "me";
-  status: "delivered" | "read";
-};
-
-export type chatsDBProps = {
-  id: number;
-  name: string;
-  avatar: string;
-  lastMessage: string;
-  lastMessageTime: Date | string | undefined;
-  unreadCount: number;
-  isOnline: boolean;
-  messages: ChatMessage[];
-};
-
-export const chatsDB: chatsDBProps[] = [
-  {
-    id: 1,
-    name: "Pablo Demarchi",
-    messages: [
-      {
-        text: "Che al final encontraste el lugar para estacionar?",
-        timestamp: "6:46 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, me costó un poco pero dejé el auto a la vuelta jeje, por suerte nadie me vió estacionando 🤣🤣🤣🤣🤣. Tuve que hacer malabares para meter el auto en ese lugarcito que dejaron jajaja",
-        timestamp: "6:47 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Bueno joya, ya entro entonces!",
-        timestamp: "6:48 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tremenda la peli eh, la verdad me re sorprendió",
-        timestamp: "9:30 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Mal! La escena donde pelean en el agua me dejó loco 😱, estaban muy buenos los efectos la verdad y encima el sonido era brutal, aunque por ahí se saturaba un poco nomás 😂😂😂😂😂, pero todo lo demas estubo súper",
-        timestamp: "9:31 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Siii, no me la esperaba ni ahí",
-        timestamp: "9:32 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Después te paso un video con teorías que vi en YouTube jajaj",
-        timestamp: "9:33 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, estaría genial eso. Estoy re manija jaja",
-        timestamp: "9:34 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ahora justamente estaba viendo a un YouTuber que hablaba de la peli",
-        timestamp: "9:34 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No le entendí mucho así que me voy a comer algo y después sigo",
-        timestamp: "9:34 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Che y mañana qué onda? Vas a lo de Nico?",
-        timestamp: "10:02 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No sé todavía, me siento medio bajón hoy la verdad",
-        timestamp: "10:04 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Uh amigo, todo bien?",
-        timestamp: "10:05 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, tranqui, solo depre de domingo jaja. Viste cómo es.",
-        timestamp: "10:06 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Te entiendo. Si querés venite a casa y comemos algo, sin plan ni nada.",
-        timestamp: "10:07 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Re va eso, gracias loco ❤️",
-        timestamp: "10:08 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Para eso estamos pa",
-        timestamp: "10:08 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Mañana te caigo tipo 8 entonces",
-        timestamp: "10:10 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, te espero con empanadas 😎",
-        timestamp: "10:11 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Jajaja sos un grande",
-        timestamp: "10:12 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Che viste lo de la tormenta?",
-        timestamp: "8:22 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, me despertó el viento 😵",
-        timestamp: "8:23 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Encima se cortó la luz un toque acá",
-        timestamp: "8:24 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Acá también. No puedo prepararme ni el café, encima me levante temprano para terminar de hacer un trabajo. y bueno... lo dejaré para después de comer 🤣🤣🤣",
-        timestamp: "8:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Iagual, creo que es mejor descansar y esperar a que pase todo",
-        timestamp: "8:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Por las dudas voy a desenchufar mis cosas jajaja",
-        timestamp: "8:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Drama real jaja",
-        timestamp: "8:25 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Creo que yo debería hacer lo mismo",
-        timestamp: "8:26 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "🤣🤣🤣🤣🤣",
-        timestamp: "8:26 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Che me olvidé de contarte, me llamaron de laburo nuevo",
-        timestamp: "3:00 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ehhh qué bien amigooo! Te felicitoooo 🥳, tenemos que salir a festejar en algún lugar amigo, encima con la dedicación que tenés vas a llegar a lejos amigo!!! Ahora le voy a decir a mamá que se va a poner súper contenta",
-        timestamp: "3:01 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Gracias loco, todavía no lo puedo creer jajaja",
-        timestamp: "3:02 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Te re lo merecés, posta",
-        timestamp: "3:03 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Vamos a tener que festejar eso eh",
-        timestamp: "3:04 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ni hablar. Este finde armamos algo.",
-        timestamp: "3:05 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Avísame y llevo birras 😏",
-        timestamp: "3:06 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0000.webp",
-    lastMessage: "",
-    lastMessageTime: new Date(Date.now() - 5 * 60 * 1000),
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 2,
-    name: "Martín Pérez",
-    messages: [
-      {
-        text: "Che, al final vas a la juntada el sábado?",
-        timestamp: "10:01 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Mirá que no podes faltar amigo vos he jajaja, hace mucho no nos vemos así que nos viene de 10",
-        timestamp: "10:01 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tenemos que ponernos al día con nuestras cosas 🤣🤣🤣🤣",
-        timestamp: "10:01 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, obvio. A qué hora era?",
-        timestamp: "10:02 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ya le avisé a la Eli que si o si voy jajaja asi que nos vemos allá seguro amigo",
-        timestamp: "10:02 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Avisame nomas la hora y caigo 🚀",
-        timestamp: "10:02 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tipo 8 voy maso, pero podés caer antes si querés así habalmos.",
-        timestamp: "10:03 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, capaz llego 7 y media así ayudo con algo.",
-        timestamp: "10:04 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Buenísimo, traé la Play si podés de paso 🤣🤣🤣🤣.",
-        timestamp: "10:05 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "La mia dejó de andar hace unos 2 mesesmaso 😑😑😑😑",
-        timestamp: "10:05 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Obvio, llevo los joysticks también.",
-        timestamp: "10:06 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tengo un juego de cartas que está buenisimo, ese tambien voy a tratar de llevar",
-        timestamp: "10:06 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Voy a pedirle a la Eli que lleve alargue para enchufar todo nomás 🤣🤣🤣🤣",
-        timestamp: "10:06 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Che, quién más va?",
-        timestamp: "10:07 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dalee perfecto, me parece genial jajajaja. Viene el Nico, la Vale, el Fede y capaz la Sofi.",
-        timestamp: "10:08 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Despues otros amigos de Eli que no se quienes son jajajaja",
-        timestamp: "10:08 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Buee la Sofiii. Hace mil que no la veo jajaja.",
-        timestamp: "10:09 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, encima dice que anda con ganas de salir jajajaja 🤣🤣.",
-        timestamp: "10:10 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Che, y no sabes si hay que llevar algo mas para tomar?",
-        timestamp: "10:11 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Lo que vos quieras tomas nomás, acá creeria que va a haber algo, pero no sé que.",
-        timestamp: "10:12 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, paso por el chino antes de ir entonces y compro.",
-        timestamp: "10:13 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Hace falta hielo tambien?",
-        timestamp: "10:14 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tienen muy buenas ofertas ahora así que llevo",
-        timestamp: "10:14 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, si querés traete un par de bolsas por las dudas.",
-        timestamp: "10:15 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Listo, anoto todo. Algo más?",
-        timestamp: "10:16 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No, yo creo que con eso estamos. Nos vemos el sábado amigazo!",
-        timestamp: "10:17 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale mi hermano, cualquier cosa te estoy chiflando",
-        timestamp: "10:18 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Espero no me pase lo de la play 🤣🤣🤣🤣",
-        timestamp: "10:18 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Mientras tanto voy preparando el juego de cartas, está buenísimo, tenes que formar dos grupos y vos tenés que tratar de hacer que tu equipo adivine la mayor cantidad de cartas en el menor tiempo posible. En la carta, abajo aparecen palabras que no podes decir. Está muy bueno!!",
-        timestamp: "10:18 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Que copado amigo, suena re interesante jajaja. Che tenés la direccion del lugar? sino te paso.",
-        timestamp: "10:19 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, la tengo pero pasala igual así la guardo.",
-        timestamp: "10:20 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Es en la casa de siempre, la de la esquina con el portón verde.",
-        timestamp: "10:21 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Perfecto, ya sé cuál es 🙂.",
-        timestamp: "10:22 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Vas a venir con alguien?",
-        timestamp: "10:23 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Capaz vaya con mi primo Ale,",
-        timestamp: "10:24 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, de una, que venga si jajajaja.",
-        timestamp: "10:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0002.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 3,
-    name: "Andrea Gómez",
-    messages: [
-      {
-        text: "Nos vemos en el café tipo 5?",
-        timestamp: "3:10 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, ahí nos vemos.",
-        timestamp: "3:12 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Querés que lleve algo?",
-        timestamp: "3:13 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No hace falta, tranqui.",
-        timestamp: "3:14 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Bueno, igual llevo unas medialunas por las dudas jaja",
-        timestamp: "3:15 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Jajaja, sos un crack",
-        timestamp: "3:16 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Te acordás dónde queda el café, no?",
-        timestamp: "3:17 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, el de siempre, ¿no?",
-        timestamp: "3:18 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ese mismo. Igual te paso la ubicación por las dudas.",
-        timestamp: "3:19 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, mandala así la guardo.",
-        timestamp: "3:20 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Listo, te la mandé por WhatsApp.",
-        timestamp: "3:21 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Perfecto, ya la tengo.",
-        timestamp: "3:22 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Vas a ir en auto o en bondi?",
-        timestamp: "3:23 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Voy en auto, así después te llevo si querés.",
-        timestamp: "3:24 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Buenísimo, gracias!",
-        timestamp: "3:25 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Querés que pase a buscarte?",
-        timestamp: "3:26 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No, tranqui, voy directo. Nos vemos allá!",
-        timestamp: "3:27 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0003.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 4,
-    name: "Alberto Torres",
-    messages: [
-      {
-        text: "Te viste el último capi de la serie?",
-        timestamp: "7:15 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí! Altísimo final.",
-        timestamp: "7:20 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No me esperaba ese giro, posta.",
-        timestamp: "7:21 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Yo tampoco, me dejó re manija.",
-        timestamp: "7:22 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Cuándo sale el próximo?",
-        timestamp: "7:23 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Creo que el viernes que viene.",
-        timestamp: "7:24 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Uff, falta una banda.",
-        timestamp: "7:25 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Encima vi teorías re locas en Twitter.",
-        timestamp: "7:26 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Pasame alguna, así me cebás más.",
-        timestamp: "7:27 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Después te paso el hilo, está buenísimo.",
-        timestamp: "7:28 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, lo leo después de cenar.",
-        timestamp: "7:29 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Ya comiste?",
-        timestamp: "7:30 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Todavía no, estoy viendo qué hago.",
-        timestamp: "7:31 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Pedite una pizza, es la que va.",
-        timestamp: "7:32 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Me tentaste, ahora llamo a la pizzería.",
-        timestamp: "7:33 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Jajaja, después contame si estaba buena.",
-        timestamp: "7:34 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Obvio, te mando foto.",
-        timestamp: "7:35 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0004.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 5,
-    name: "Franco Fernández",
-    messages: [
-      {
-        text: "Che, necesito una mano con el proyecto.",
-        timestamp: "11:00 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Decime nomás, en qué te doy una mano?",
-        timestamp: "11:02 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No entiendo bien cómo hacer el login.",
-        timestamp: "11:03 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Tenés el repo a mano?",
-        timestamp: "11:04 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, te lo paso por mail.",
-        timestamp: "11:05 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, avisame cuando lo mandes.",
-        timestamp: "11:06 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Listo, ya te lo mandé.",
-        timestamp: "11:07 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Perfecto, ahora lo miro.",
-        timestamp: "11:08 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Te paso mi usuario y clave de prueba?",
-        timestamp: "11:09 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, así pruebo bien.",
-        timestamp: "11:10 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Usuario: franco, clave: 1234",
-        timestamp: "11:11 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Listo, ya entré.",
-        timestamp: "11:12 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "¿Te tira algún error?",
-        timestamp: "11:13 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No, pero el botón de login no se desactiva.",
-        timestamp: "11:14 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ah, seguro es por el estado. Fijate si lo actualizás bien.",
-        timestamp: "11:15 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, ahora lo reviso.",
-        timestamp: "11:16 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Cualquier cosa avisame, estoy en la compu.",
-        timestamp: "11:17 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0005.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 6,
-    name: "Jorge Martínez",
-    messages: [
-      {
-        text: "Hola Jorge, me pasás el apunte de historia?",
-        timestamp: "8:30 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, bancá que lo busco.",
-        timestamp: "8:31 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Gracias! Me re salvas.",
-        timestamp: "8:32 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "No hay drama, para eso estamos.",
-        timestamp: "8:33 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Te lo mando por mail o por acá?",
-        timestamp: "8:34 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Por acá mejor, así lo bajo al toque.",
-        timestamp: "8:35 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Listo, ahí va el PDF.",
-        timestamp: "8:36 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Llegó perfecto, mil gracias.",
-        timestamp: "8:37 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ya estudiaste algo?",
-        timestamp: "8:38 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Un poco, pero me falta banda todavía.",
-        timestamp: "8:39 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Tranqui, yo tampoco arranqué fuerte.",
-        timestamp: "8:40 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Querés que repasemos juntos después?",
-        timestamp: "8:41 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, de una. Hoy a la tarde podés?",
-        timestamp: "8:42 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, tipo 5 te va?",
-        timestamp: "8:43 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Perfecto, te escribo cuando esté libre.",
-        timestamp: "8:44 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale, suerte con el estudio!",
-        timestamp: "8:45 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Gracias, igualmente!",
-        timestamp: "8:46 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0006.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 7,
-    name: "Sofía Ramírez",
-    messages: [
-      {
-        text: "Feliz cumple 🎉",
-        timestamp: "7:55 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Gracias genia! 🥳",
-        timestamp: "7:58 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0007.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 8,
-    name: "Marco Perdomo",
-    messages: [
-      {
-        text: "Buen día, revisaste el informe?",
-        timestamp: "9:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Ey Marco, puedo pasar por tu casa a buscar los apuntes?",
-        timestamp: "9:28 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0008.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 9,
-    name: "Paula Lezcano",
-    messages: [
-      {
-        text: "Buen día, miraste el informe?",
-        timestamp: "9:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Hola Pau, sigue estando para pasar a buscarlo?",
-        timestamp: "9:28 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0009.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 10,
-    name: "Pedro Falcón",
-    messages: [
-      {
-        text: "Buen día, revisaste el informe?",
-        timestamp: "9:25 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Todavía no pude, pero mañana sin falta te lo paso.",
-        timestamp: "9:28 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0010.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 11,
-    name: "Lucía Méndez",
-    messages: [
-      {
-        text: "Te llegó el paquete al final?",
-        timestamp: "8:00 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, todo perfecto. Gracias che.",
-        timestamp: "8:05 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0011.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 12,
-    name: "Diego Rivera",
-    messages: [
-      {
-        text: "Tenemos la reunión a las 3 eh.",
-        timestamp: "11:15 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale! Voy a estar al toque.",
-        timestamp: "11:17 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0012.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 13,
-    name: "Valeria Suárez",
-    messages: [
-      {
-        text: "Me pasás el archivo?",
-        timestamp: "10:20 a.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Obvio! Te lo mando ya.",
-        timestamp: "10:21 a.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0013.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 14,
-    name: "Camila Quintana",
-    messages: [
-      {
-        text: "Jugamos esta noche?",
-        timestamp: "6:30 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Dale! Después de cenar me conecto.",
-        timestamp: "6:33 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0014.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-  {
-    id: 15,
-    name: "Matías Ortega",
-    messages: [
-      {
-        text: "Viste el correo de RRHH?",
-        timestamp: "4:45 p.m.",
-        sender: "contact",
-        status: "delivered",
-        id: 0,
-      },
-      {
-        text: "Sí, ya les respondí. Vos qué onda?",
-        timestamp: "4:47 p.m.",
-        sender: "me",
-        status: "delivered",
-        id: 0,
-      },
-    ],
-    avatar: "/Avatars/image_0015.webp",
-    lastMessage: "",
-    lastMessageTime: undefined,
-    unreadCount: 0,
-    isOnline: false,
-  },
-];
